Add a basic double-chained list implementation.
The implementation is designed to work with ARC activated. Nodes have a pointer to the previous and the next entry in the list but don't retain those pointers. In addition, it takes a second pointer to the next entry whenever this one does not point to the root of the list. That last pointer is retained, ensuring list integrity without cross-retaining issues.