Major overhaul of the command workflow.
Introduce a notion of command group. A command group allow several commands to be queued at once while ensuring that a failure in one of the command will abort the whole group. This allow writing code in a way that looks much more "synchronous" and dramatically reduce the need of command queuing within a onDone helper.

onDone helper now return a BOOL value that mean:
* YES: I handled the result
* NO: I didn't handle the result, it must be handled by my parent group (and so one)

This allow error propagation through the group layer, when the top-level group receive message, it triggers a disconnection.